refactored: centralize content store tests to a single test suite
