+++ src/stores/content.ts
@@ -1,4 +1,7 @@
+import { defineStore } from "pinia";
+// @ts-ignore
 import { useInMemoryContentStore } from "./inMemoryContent";
+// @ts-ignore
 import { useLocalStorageContentStore } from "./localStorageContent";
 
 export const useContentStore = (() => {
